// src/services/types/useCase/product_roadmapping.ts
export type ProductRoadmappingApiResponse = {
  _id: string
  product_roadmapping: {
    section1: {
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      text1_en: string
      text1_fa: string
      button1_en: string
      button1_fa: string
    }
    section2: {
      title1_en: string
      title1_fa: string
      text1_en: string
      text1_fa: string
      image1: string
    }
    section3: {
      title1_en: string
      title1_fa: string
      text1_en: string
      text1_fa: string
      button1_en: string
      button2_fa: string
      image1: string
    }
    section4: {
      image1: string
      image2: string
      image3: string
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      title3_en: string
      title3_fa: string
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      text3_en: string
      text3_fa: string
    }
    section5: {
      title1_en: string
      title1_fa: string
      text1_en: string
      text1_fa: string
      image1: string
    }
    section6: {
      title1_en: string
      title1_fa: string
      button1_en: string
      button1_fa: string
    }
    section7: {
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      button1_en: string
      button1_fa: string
    }
  }
  __v: number
}